added new

+++ question16.c
@@ -2,15 +2,15 @@
 #include<conio.h>
 void main()
 {
-	float hrs,min,sec;
+	int hrs,min,sec;
 	printf("Enter your time in hours:");
-	scanf("%f",&hrs);
+	scanf("%d",&hrs);
 	
 	min=hrs*60;
-	printf("The conversion of hours into minute is %.2f min.",min);
+	printf("The conversion of hours into minute is %.2d min.",min);
 	
 	sec=min*60;
-	printf("\nThe conversion of minute into sec is %.2f min.",sec);
+	printf("\nThe conversion of minute into sec is %.2d min.",sec);
 	
 	return 0;
